% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_area_details}
\alias{cvd_area_details}
\title{Area details}
\usage{
cvd_area_details(time_period_id = 1, area_id = 1)
}
\arguments{
\item{time_period_id}{integer - specified time period for which to return details for, i.e. population and participation rate (compulsory)}

\item{area_id}{integer - specified area id for which to return details for.}
}
\value{
Named list of tibbles containing area 'area_details', child 'area_child_details' (where appropriate) and parent 'area_parent_details' (where appropriate)
}
\description{
Returns details of a specific area at a given time period, including details about any parent and child areas.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/area/<area_id>/details]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Farea\%2F\%3Carea_id\%3E\%2Fdetails}
}
\examples{
area_details <- cvd_area_details(area_id = 1, time_period_id = 17)
}
