% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_area_flat_subsystems}
\alias{cvd_area_flat_subsystems}
\title{Area flat subsystems}
\usage{
cvd_area_flat_subsystems(area_id = 5)
}
\arguments{
\item{area_id}{integer - the area to return data for (compulsory)}
}
\value{
tibble of data
}
\description{
Similar to \code{cvd_area_nested_subsystems()} but the sub-areas are grouped
based on their system level.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/area/<area_id>/flatSubSystems]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Farea\%2F\%3Carea_id\%3E\%2FflatSubSystems}
}
\examples{
test <- cvd_area_flat_subsystems(area_id = 5)
}
