% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_area_search}
\alias{cvd_area_search}
\title{Search for matching areas}
\usage{
cvd_area_search(partial_area_name = "Surgery", time_period_id = 1)
}
\arguments{
\item{partial_area_name}{string - string to use to search for an Area (compulsory)}

\item{time_period_id}{integer - limits the search to Areas which have data in specified time period (compulsory)}
}
\value{
Tibble of matching areas
}
\description{
Returns a list of Areas that match a partial name for a given time period.
Uses simple LIKE '\%<partial_area_name>\%' comparison.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/area/search]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Farea\%2Fsearch}
}
\examples{
test <- cvd_area_search(partial_area_name = 'Station', time_period_id = 17)
}
