% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_details}
\alias{cvd_indicator_details}
\title{Indicator details}
\usage{
cvd_indicator_details(indicator_id = 1)
}
\arguments{
\item{indicator_id}{integer - the ID for the indicator (compulsory)}
}
\value{
tibble
}
\description{
Returns details of a single indicator
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/<indicator_ID>/details]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2F\%3Cindicator_ID\%3E\%2Fdetails}
}
\examples{
test <- cvd_indicator_details()
test <- cvd_indicator_details(indicator_id = 7)
}
