% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_list}
\alias{cvd_indicator_list}
\title{List indicators}
\usage{
cvd_indicator_list(time_period_id = 1, system_level_id = 2)
}
\arguments{
\item{time_period_id}{integer - time period to reutrn data for (compulsory)}

\item{system_level_id}{integer - system level to return data for (compulsory)}
}
\value{
tibble of time period details
}
\description{
Returns basic details of all indicators for a given system level and time period.
Only returns indicators for which data exists in selected time period, and on
selected system level. Used to populate available indicator list in Data Explorer.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/list]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2Flist}
}
\examples{
cvd_indicators <- cvd_indicator_list()
}
