% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_metric_systemlevel_comparison}
\alias{cvd_indicator_metric_systemlevel_comparison}
\title{Indicator metric system level comparison}
\usage{
cvd_indicator_metric_systemlevel_comparison(
  metric_id = 1,
  time_period_id = 1,
  area_id = 50
)
}
\arguments{
\item{metric_id}{integer - the metric to return data for (compulsory)}

\item{time_period_id}{integer - the time period to return data for (compulsory)}

\item{area_id}{integer - the area to return data for (compulsory)}
}
\value{
tibble of data
}
\description{
Returns data for the SystemLevel Comparison chart for provided metric, area
and time period. \code{Data} contains the target value as well as an array
\code{SystemLevels} which contains data grouped by system level.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/metricSystemLevelComparison/<metric_ID>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FmetricSystemLevelComparison\%2F\%3Cmetric_ID\%3E}
}
\examples{
test <- cvd_indicator_metric_systemlevel_comparison()
}
