% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_metric_timeseries}
\alias{cvd_indicator_metric_timeseries}
\title{Indicator time series by metric}
\usage{
cvd_indicator_metric_timeseries(metric_id = 1, area_id = 50)
}
\arguments{
\item{metric_id}{integer - the metric to return data for (compulsory)}

\item{area_id}{integer - the area to return data for (compulsory)}
}
\value{
tibble
}
\description{
Returns data for the time series chart for specified metric ID and area ID.
Contains an array of two areas in \code{Areas}, one of which is the National data
with the other corresponding to the provided area ID. \code{TargetValue} is also
returned in the \code{Data} dictionary.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/timeSeriesByMetric/<metric_ID>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FtimeSeriesByMetric\%2F\%3Cmetric_ID\%3E}
}
\examples{
test <- cvd_indicator_metric_timeseries()
}
