% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_nationalarea_metric_data}
\alias{cvd_indicator_nationalarea_metric_data}
\title{Indicator national vs area metric data}
\usage{
cvd_indicator_nationalarea_metric_data(
  metric_id = 1,
  time_period_id = 17,
  area_id = 739
)
}
\arguments{
\item{metric_id}{integer - metric for which to return data (compulsory)}

\item{time_period_id}{integer - time period for which to return data (compulsory)}

\item{area_id}{integer - area for which to return data (compulsory)}
}
\value{
tibble of data
}
\description{
Returns national and area data for provided metric, area and time period.
Target data contains the target value as a percentage stored as whole number
up to 100; target patients is the number of patients more needed to reach
the target percentage. If there is not data for both national and chosen
area an error will be returned.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/nationalVsAreaMetricData/<metric_ID>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FnationalVsAreaMetricData\%2F\%3Cmetric_ID\%3E}
}
\examples{
test <- cvd_indicator_nationalarea_metric_data()
}
