% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_pathway_group}
\alias{cvd_indicator_pathway_group}
\title{Pathway groups}
\usage{
cvd_indicator_pathway_group(pathway_group_id = 10)
}
\arguments{
\item{pathway_group_id}{integer - the pathway to return data for (compulsory)}
}
\value{
tibble of data
}
\description{
Pathway groups are sub-groupings of Priority Groups, visible in the Regional
& ICS Insights page. This endpoint returns a single pathway group for a given
group ID. An error will be returned if there is no pathway group associated
with the given ID. For a valid request, Pathway Group ID and named are
returned as key value pairs and the Indicators populate an array.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/pathwayGroup/<pathway_group_id>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FpathwayGroup\%2F\%3Cpathway_group_id\%3E}
}
\examples{
test <- cvd_indicator_pathway_group()
}
