% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_priority_groups}
\alias{cvd_indicator_priority_groups}
\title{Indicator priority groups}
\usage{
cvd_indicator_priority_groups()
}
\value{
tibble of data
}
\description{
Returns the list of top-level groupings (Priority Groups) displayed in the
Regional & ICS Insights page. Returns a dictionary called 'PriorityGroups'
with each key being a Priority Group name, and each value being the array of
indicators contained in that group. The 'PriorityGroupDisplayOrder'
indicates the order in whcih it should be displayed for the given Priority
Group.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/priorityGroups]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FpriorityGroups}
}
\examples{
test <- cvd_indicator_priority_groups()
}
