% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_sibling}
\alias{cvd_indicator_sibling}
\title{Indicator sibling data}
\usage{
cvd_indicator_sibling(time_period_id = 17, area_id = 30, metric_id = 1)
}
\arguments{
\item{time_period_id}{integer - time period for which to return data (compulsory)}

\item{area_id}{integer - area for which all sibling data will be returned (compulsory)}

\item{metric_id}{integer - metric for which to return data (compulsory)}
}
\value{
tibble of data
}
\description{
Returns all sibling areas and their data for specified time period, area
and metric. This endpoint is intended to only return data for selected metric,
and not all metrics for a chosen indicator, hence the metric_id query parameter.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/siblingData]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FsiblingData}
}
\examples{
est <- cvd_indicator_sibling()
}
