% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_tags}
\alias{cvd_indicator_tags}
\title{Indicator tags}
\usage{
cvd_indicator_tags()
}
\value{
tibble of tags
}
\description{
Returns a list of all available tags, which can be used to filter indicators.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/tags]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2Ftags}
}
\examples{
test <- cvd_indicator_tags()
}
