% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_external_resource}
\alias{cvd_external_resource}
\title{External resource}
\usage{
cvd_external_resource()
}
\value{
tibble fo data
}
\description{
Returns a list of all external resources
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/externalResource]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2FexternalResource}
}
\examples{
test <- cvd_external_resource()
}
