% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator}
\alias{cvd_indicator}
\title{Indicators}
\usage{
cvd_indicator(time_period_id = 1, area_id = 1, tag_id)
}
\arguments{
\item{time_period_id}{integer - time period to return data for (compulsory)}

\item{area_id}{integer - area to return data for (compulsory)}

\item{tag_id}{numeric vector - allows filtering indicators by one or more tags (optional, array)}
}
\value{
list of tibbles
}
\description{
Returns all indicators and data for a given time period and area. Also returns
time series data for all time periods available. If tags are specified, only
indicators which have one of the specified tags will be returned.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator}
}
\examples{
test <- cvd_indicator()
test <- cvd_indicator(time_period_id = 17, area_id = 55, tag_id = c(12, 13))
}
