% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_data}
\alias{cvd_indicator_data}
\title{Indicator data}
\usage{
cvd_indicator_data(indicator_id = 2, time_period_id = 1, area_id = 2)
}
\arguments{
\item{indicator_id}{integer - indicator for which to return data (compulsory)}

\item{time_period_id}{integer - time period for which to return data for (compulsory)}

\item{area_id}{integer - area for which to return data for (compulsory)}
}
\value{
tibble of indicator data
}
\description{
Returns all metric data for a specified indicator. Data will include values
for both selected area, and organisation at National Level (usually England).
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/<indicator_id>/data]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2F\%3Cindicator_id\%3E\%2Fdata}
}
\examples{
test <- cvd_indicator_data()
}
