% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_group}
\alias{cvd_indicator_group}
\title{Indicator group}
\usage{
cvd_indicator_group(indicator_group_id = 15)
}
\arguments{
\item{indicator_group_id}{integer - the group to return data for (compulsory)}
}
\value{
tibble of data
}
\description{
Returns a single indicator group for a given group ID. An error will be
returned if there is no indicator group associated with the given ID.
\code{IndicatorGroup} is the primary key in the IndicatorGroup table, which also
contains \code{IndicatorGroupName} and \code{IndicatorGroupTypeID}. The group type ID
tells you what type of indicator group you're dealing with, e.g. a Priority
Group.
\code{IndicatorGroupTypeID} is the primary key of IndicatorGroupType and so
\code{IndicatorGroupTypeName} is the associated name for the given group type ID.
Finally, there is the array of indicators which are contained in this group,
including display orders for the given group.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/indicatorGroup/<indicator_group_ID>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FindicatorGroup\%2F\%3Cindicator_group_ID\%3E}
}
\examples{
test <- cvd_indicator_group()
}
