% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_metric_data}
\alias{cvd_indicator_metric_data}
\title{Metric data}
\usage{
cvd_indicator_metric_data(metric_id = 7, time_period_id = 1, area_id = 2)
}
\arguments{
\item{metric_id}{integer - metric for which to return data for (compulsory)}

\item{time_period_id}{integer - time period for which to return data for (compulsory)}

\item{area_id}{integer - area for which to return data for (compulsory)}
}
\value{
tibble of metric data
}
\description{
Returns all metric data for a specified metric. Data will include values
for both selected area and organisation at National Level (usually England).
}
\details{
CVD Prevent API documentation:
\code{\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#*Proposed*/indicator/metric/<metric_id>/data]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#*Proposed*\%2Findicator\%2Fmetric\%2F\%3Cmetric_id\%3E\%2Fdata}}
}
\examples{
test <- cvd_indicator_metric_data()
}
