% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_metric_list}
\alias{cvd_indicator_metric_list}
\title{List metrics for indicators}
\usage{
cvd_indicator_metric_list(time_period_id = 1, system_level_id = 1)
}
\arguments{
\item{time_period_id}{integer - time period to return data for (compulsory)}

\item{system_level_id}{integer - system level to return data for (compulsory)}
}
\value{
tibble of indicators left-joined with metrics
}
\description{
Returns same data as cvd_indicator_list() but adds a 'MetricList' array for
each indicator, containing details of the relevant metrics. Only returns
indicators for which data exists in selected time period, and on selected
system level.
}
\details{
CVD Prevent API documentation:
\code{\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#*Proposed*/indicator/metricList]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#*Proposed*\%2Findicator\%2FmetricList}}
}
\examples{
test <- cvd_indicator_metric_list(time_period_id = 1, system_level_id = 1)
}
