% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_person_timeseries}
\alias{cvd_indicator_person_timeseries}
\title{Indicator persons time series by indicator}
\usage{
cvd_indicator_person_timeseries(indicator_id = 1, area_id = 1)
}
\arguments{
\item{indicator_id}{integer - the indicator to return data for (compulsory)}

\item{area_id}{integer - the area to return data for (compulsory)}
}
\value{
tibble of data
}
\description{
Returns data for the Inequalities Markers Time Series chart for the provided
indicator ID and area ID. \code{Data} contains information about the chosen
target value as well as an array \code{InequalityMarkers} which contains all the
time series data grouped into metric category types e.g. age group,
ethnicity, etc.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/personsTimeSeriesByIndicator/<indicator_ID>]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FpersonsTimeSeriesByIndicator\%2F\%3Cindicator_ID\%3E}
}
\examples{
test <- cvd_indicator_person_timeseries()
}
