% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_raw_data}
\alias{cvd_indicator_raw_data}
\title{Retrieve raw metric values for multiple metrics, a specified are and time period}
\usage{
cvd_indicator_raw_data(indicator_id, time_period_id, system_level_id)
}
\arguments{
\item{time_period_id}{Integer (required). The reporting period (time period) for which to retrieve data. Use \code{\link[=cvd_time_period_list]{cvd_time_period_list()}} to find valid IDs.}

\item{area_id}{Integer (required). The AreaID for which to retrieve data. Use \code{\link[=cvd_area_list]{cvd_area_list()}} or \code{\link[=cvd_area_search]{cvd_area_search()}} to find valid IDs.}

\item{metric_id}{Integer vector (required). One or more MetricIDs specifying which metrics to return. Use \code{\link[=cvd_indicator_metric_list]{cvd_indicator_metric_list()}} or \code{\link[=cvd_indicator_data]{cvd_indicator_data()}} to find valid MetricIDs.}
}
\value{
A tibble with one row per metric breakdown for all requested metrics, the specified area and period. Typical columns include \code{MetricID}, \code{AreaID}, \code{Value}, \code{Numerator}, \code{Denominator}, \code{LowerConfidenceLimit}, \code{UpperConfidenceLimit}, \code{MetricCategoryName}, \code{CategoryAttribute}, and others.
}
\description{
Returns raw values for multiple metrics (specified as a vector of \code{metric_id}) for a single NHS area and reporting period using the CVDPREVENT API. This function fetches unfiltered raw data at the metric level, allowing comprehensive extraction for all selected metrics and their available breakdowns (such as by age, sex or other category) within the chosen context.
}
\details{
Use this function to retrieve a wide set of metric breakdowns for multiple metrics in a single area and time period - useful for broad data extractions, dashboards or advanced analytics.
}
\section{API Documentation}{

See the \href{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2FrawData}{CVDPREVENT API documentation: Indicator raw data}
}

\examples{
# Retrieve raw data for metrics 126 and 127 in area 1103 for time period 17
cvd_indicator_raw_data(time_period_id = 17, area_id = 1103, metric_id = c(126, 127)) |>
  dplyr::select(MetricID, AreaID, Value, MetricCategoryName, CategoryAttribute)

# Find a vector of valid metric IDs, then retrieve raw data for one area
metrics <- cvd_indicator_metric_list(time_period_id = 17, system_level_id = 5)
cvd_indicator_raw_data(time_period_id = 17, area_id = 1103, metric_id = metrics$MetricID[1:3])

}
