% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_raw_data}
\alias{cvd_indicator_raw_data}
\title{Indicator raw data (JSON)}
\usage{
cvd_indicator_raw_data(
  indicator_id = 1,
  time_period_id = 1,
  system_level_id = 1
)
}
\arguments{
\item{indicator_id}{integer - indicator for which to return data for (compulsory)}

\item{time_period_id}{integer - time period for which to return data for (compulsory)}

\item{system_level_id}{integer - system level for which to return data for (compulsory)}
}
\value{
tibble of indicator data
}
\description{
Returns all metric data for a specified indicator, system level and time
period.
}
\details{
CVD Prevent API documentation:
\link[=https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#/indicator/<indicator_ID>/rawDataJSON]{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2Findicator\%2F\%3Cindicator_ID\%3E\%2FrawDataJSON}
}
\examples{
test <- cvd_indicator_raw_data()
}
