% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_indicator_types}
\alias{cvd_indicator_types}
\title{List indicator types}
\usage{
cvd_indicator_types()
}
\value{
Tibble of indicator types
}
\description{
Returns IDs and descriptions for indicator types.
This is a helper function for the \code{cvd_time_period_list()} which permits the
optional parameter of \code{indicator_type_id}.
}
\examples{
# List available indicator types
cvd_indicator_types()
}
\seealso{
\code{\link[=cvd_time_period_list]{cvd_time_period_list()}}
}
