% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_api_functions.R
\name{cvd_time_period_list}
\alias{cvd_time_period_list}
\title{List time periods}
\usage{
cvd_time_period_list(opt_indicator_type_id)
}
\arguments{
\item{opt_indicator_type_id}{integer - Indicator type ID, e.g. standard or outcome indicator type. If passed will show time periods containing data of the given type (optional)}
}
\value{
Tibble of time period details
}
\description{
Returns all available time periods
}
\details{
CVD Prevent API documentation:
\href{https://bmchealthdocs.atlassian.net/wiki/spaces/CP/pages/317882369/CVDPREVENT+API+Documentation#\%2FtimePeriod}{Time period}
}
\examples{
cvd_time_periods <- cvd_time_period_list()
}
